fix: use uint64 for interface statistics counters
The /proc/net/dev as well as /sys/class/net/{iface}/statistics uses struct
rtnl_link_stats64, where all counters are __u64.

Previously, parsing these values into int64 caused errors for large counters:

    strconv.ParseInt: parsing "18446744073709551613": value out of range

This commit ensures that all counters are treated as uint64.

See: https://www.kernel.org/doc/html/latest/networking/statistics.html

+++ lxd-agent/metrics.go
@@ -390,14 +390,14 @@ func getNetworkMetrics() (map[string]metrics.NetworkMetrics, error) {
 	for dev, state := range networkState() {
 		stats := metrics.NetworkMetrics{}
 
-		stats.ReceiveBytes = uint64(state.Counters.BytesReceived)
-		stats.ReceiveDrop = uint64(state.Counters.PacketsDroppedInbound)
-		stats.ReceiveErrors = uint64(state.Counters.ErrorsReceived)
-		stats.ReceivePackets = uint64(state.Counters.PacketsReceived)
-		stats.TransmitBytes = uint64(state.Counters.BytesSent)
-		stats.TransmitDrop = uint64(state.Counters.PacketsDroppedOutbound)
-		stats.TransmitErrors = uint64(state.Counters.ErrorsSent)
-		stats.TransmitPackets = uint64(state.Counters.PacketsSent)
+		stats.ReceiveBytes = state.Counters.BytesReceived
+		stats.ReceiveDrop = state.Counters.PacketsDroppedInbound
+		stats.ReceiveErrors = state.Counters.ErrorsReceived
+		stats.ReceivePackets = state.Counters.PacketsReceived
+		stats.TransmitBytes = state.Counters.BytesSent
+		stats.TransmitDrop = state.Counters.PacketsDroppedOutbound
+		stats.TransmitErrors = state.Counters.ErrorsSent
+		stats.TransmitPackets = state.Counters.PacketsSent
 
 		out[dev] = stats
 	}
